% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dst_find_val_id.R
\name{dst_find_val_id}
\alias{dst_find_val_id}
\title{This is a helper function to extract the ID's in the values list
that the dst_meta function returns.}
\usage{
dst_find_val_id(meta_data, variable, values_text = NULL)
}
\arguments{
\item{meta_data}{The meta data object.}

\item{variable}{The variable to search in.}

\item{values}{Character vector. The values you want to extract the IDs for. If NULL, the returned value will be "*".}
}
\description{
This is a helper function to extract the ID's in the values list
that the dst_meta function returns.
}

