% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_get_all_data.R
\name{dst_get_all_data}
\alias{dst_get_all_data}
\title{This function returns the entire table requested data from the StatBank.
(http://www.statistikbanken.dk/statbank5a/ or http://www.dst.dk)}
\usage{
dst_get_all_data(table, lang = "da")
}
\arguments{
\item{table}{Table from StatBank.}

\item{lang}{language. "en" for english or "da" for danish.}
}
\description{
Get all data from a table in the StatBank.
}
\seealso{
Other Data retrival functions: 
\code{\link{dst_get_data}()}
}
\author{
Aleksander Bang-Larsen <contact@aleksanderbl.dk>
}
\concept{Data retrival functions}
