% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dst_get_tables.R
\name{dst_get_tables}
\alias{dst_get_tables}
\title{This functions downloads all the available tables.}
\usage{
dst_get_tables(lang = "da")
}
\arguments{
\item{lang}{"da" (danish) or "en" (english)}
}
\description{
This functions downloads all the available tables.
}

