% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dst_meta}
\alias{dst_meta}
\title{This function returns meta data for a table from StatBank Denmark / Statistics Denmark.
(http://www.statistikbanken.dk/statbank5a/ or http://www.dst.dk)}
\usage{
dst_meta(table, ..., lang = "en")
}
\arguments{
\item{table}{The name of the table you want meta data for.}

\item{...}{Ignored.}

\item{lang}{You can choose "en" for english or "da" for danish.}
}
\description{
This function POSTs a request for meta data on a table from Statistics Denmark
and returns a JSON object with the information.
}

