% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_query_match.R
\name{dst_query_match}
\alias{dst_query_match}
\title{Helper function to return ids based on text values}
\usage{
dst_query_match(table, lang, meta_data, query, format)
}
\arguments{
\item{table}{Table from StatBank.}

\item{lang}{language. "en" for english and "da" for danish.}
}
\description{
This is a helper function to return the ids based on the text values.
}
