% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_value_limit.R
\name{dst_value_limit}
\alias{dst_value_limit}
\title{This is a helper function that returns the number
of values the call will return.
The API has a limit of 100.000 values.}
\usage{
dst_value_limit(query, dst_meta)
}
\arguments{
\item{query}{Query object to analyse}

\item{dst_meta}{Meta data to filer query with}
}
\description{
This is a helper function that returns the number
of values the call will return.
The API has a limit of 100.000 values.
}
