% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_tables.R
\docType{data}
\name{tables_en}
\alias{tables_en}
\title{Cached list of all available tables}
\format{
## `tables_en`
A data frame with 2339 rows and 8 columns:
\describe{
 \item{id}{Unique ID of the table}
 \item{text}{Description of the table}
 \item{unit}{The unit of the values represented in the table}
 \item{updated}{When this table was last updated}
 \item{firstPeriod}{The time of the first observation in the table}
 \item{latestPeriod}{The time of the latest observation in the table}
 \item{active}{Binary indicator of the table being active}
 \item{variables}{List of the variables in the table}
}
}
\usage{
tables_en
}
\description{
Cached list of all available tables
}
\keyword{datasets}
