% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_get_all_data.R
\name{dst_get_all_data}
\alias{dst_get_all_data}
\title{This function returns the entire table requested data from the StatBank.
(http://www.statistikbanken.dk/statbank5a/ or http://www.dst.dk)}
\usage{
dst_get_all_data(table, lang = "da", parse_dst_tid = TRUE)
}
\arguments{
\item{table}{Table from StatBank.}

\item{lang}{language. "en" for english or "da" for danish.}

\item{parse_dst_tid}{Logical. Default is TRUE. The date will be the first day
of the period and the time zone is "UTC" and not the Danish "CET".}
}
\description{
Get all data from a table in the StatBank. This function
provides all the query parameters on the user's behalf. You, as a user, do
not need to specify anything other than a table name and you will be given
the entire contents of the table in a nice long format. This is useful for
you, if you would like to filter the table with e.g. \code{{dplyr}} functions or
save the entire table for archival.
}
\seealso{
Other Data retrival functions: 
\code{\link{dst_get_data}()}
}
\author{
Aleksander Bang-Larsen \href{mailto:contact@aleksanderbl.dk}{contact@aleksanderbl.dk}
}
\concept{Data retrival functions}
