% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dst_get_data.R
\name{dst_get_data}
\alias{dst_get_data}
\title{This function returns the requested data from the StatBank table.
(http://www.statistikbanken.dk/statbank5a/ or http://www.dst.dk)}
\usage{
dst_get_data(table, ..., query, parse_dst_tid = TRUE, lang = "en",
  meta_data = NULL, format = "CSV", value_presentation = "Value")
}
\arguments{
\item{table}{Table from StatBank.}

\item{...}{Build your own query.}

\item{query}{A list object with your query.}

\item{lang}{language. "en" for english and "da" for danish.}

\item{format}{for now just csv but later JSON and more.}

\item{value_presentation}{for now, "value" or "default"}

\item{meta}{Meta data for the table. If NULL the meta data will be requested.}
}
\description{
Get data from a table in the StatBank.
}

