% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dst_get_data.R
\name{dst_get_data}
\alias{dst_get_data}
\title{This function returns the requested data from the StatBank table.
(http://www.statistikbanken.dk/statbank5a/ or http://www.dst.dk)}
\usage{
dst_get_data(table, ..., query = NULL, parse_dst_tid = TRUE, lang = "en",
  format = "CSV", value_presentation = "ValueAndCode")
}
\arguments{
\item{table}{Table from StatBank.}

\item{...}{Build your own query.}

\item{query}{A request for data structured like the basic_basic_query.}

\item{lang}{language. "en" for english and "da" for danish.}

\item{format}{for now just csv but later JSON and more.}

\item{value_presentation}{for know just ValueAndCode}
}
\description{
Get data from a table in the StatBank.
}

