% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dst_get_data}
\alias{dst_get_data}
\title{This function returns the requested data from the StatBank table.
(http://www.statistikbanken.dk/statbank5a/ or http://www.dst.dk)}
\usage{
dst_get_data(query, table, ..., lang = "en", format = "CSV",
  value_presentation = "ValueAndCode")
}
\arguments{
\item{request}{A request for data structured like the basic_basic_query.}

\item{lang}{language. "en" for english and "da" for danish.}

\item{format}{for now just csv but later JSON and more.}

\item{value_presentation}{for know just default}
}
\description{
Get data from a table.
}

