% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dst_get_data.R
\name{dst_get_data}
\alias{dst_get_data}
\title{This function returns the requested data from the StatBank table.
(http://www.statistikbanken.dk/statbank5a/ or http://www.dst.dk)}
\usage{
dst_get_data(table, ..., query = NULL, parse_dst_tid = TRUE, lang = "da",
  meta_data = NULL, format = "CSV", value_presentation = "Value")
}
\arguments{
\item{table}{Table from StatBank.}

\item{...}{Build your own query.}

\item{query}{A list object with your query.}

\item{parse_dst_tid}{Logical. Default is TRUE. The date will be the first
day of the period and the time zone is "UTC" and not the Danish "CET".}

\item{lang}{language. "en" for english or "da" for danish.}

\item{meta_data}{Meta data for the table. If NULL the meta data will be requested.}

\item{format}{"CSV".}

\item{value_presentation}{For now, "value" or "default"}
}
\description{
Get data from a table in the StatBank.
}

