% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dst_query_match.R
\name{dst_query_match}
\alias{dst_query_match}
\title{This is a helper function to return the ids based on the text values.}
\usage{
dst_query_match(table, lang, meta_data, query)
}
\arguments{
\item{table}{Table from StatBank.}

\item{lang}{language. "en" for english and "da" for danish.}

\item{meta_data}{}

\item{query}{}
}
\description{
This is a helper function to return the ids based on the text values.
}

